Add ClothPredictor class and update attribute categories
Added a new ClothPredictor class within the feature extraction module. This class changes the way cloth attributes are predicted, accounting for a wider variety of attributes. Related changes were made to the DeepFashion2GeneralizedCategoriesAndAttributes class to update the mask and attribute categories. Shifted from using 'Predictor' to 'ClothPredictor' in service and test files.